z3,rgd,z3ts,pcbt,symsan,tests: fmemcmp compat, replay checker, runtime regression
- z3.cpp: safe fmemcmp path copies packed op bytes instead of
  dereferencing captured values as target pointers.
- rgd-parser.cpp: replace unaligned uint64_t* cast with memcpy
  in wide constant chunking.
- z3-ts.cpp: fix uint8_t index wrap in read_concrete.
- pcbt.{hpp,cpp}: add const ReplayFullTrace for single-root
  CID/predicate/frontier consistency checks; add const
  retry_count accessor; InsertSuffix already guards frontier.
- symsan.cpp: add SYMAFL_REPLAY_CHECK diagnostic mode (full-pipe
  capture, ReplayFullTrace before insertion); add replay counters
  and log output (pcbt-replay line).
- tests/fmemcmp_runtime.cpp: runtime regression for fmemcmp
  capture flags (valid memcmp both orientations, invalid pointer
  safety via direct dfsan_union).
- eval-xz.sh: record dirty status/diffs and critical binary
  hashes in manifest.

+++ solvers/z3-ts.cpp
@@ -290,7 +290,7 @@ z3::expr Z3AstParser::read_concrete(dfsan_label label, uint16_t size) {
   }
 
   z3::expr val = context_.bv_val(itr->second[0], 8);
-  for (uint8_t i = 1; i < size; i++) {
+  for (size_t i = 1; i < size; i++) {
     val = z3::concat(context_.bv_val(itr->second[i], 8), val);
   }
   return val;
